+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#782](https://github.com/NibiruChain/nibiru/pull/782) - replace GitHub test workflows to use make commands
 * [#784](https://github.com/NibiruChain/nibiru/pull/784) - fix runsim
+* [#783](https://github.com/NibiruChain/nibiru/pull/783) - sanitise inputs for msg swap simulations
 
 ## [v0.11.0](https://github.com/NibiruChain/nibiru/releases/tag/v0.11.0) - 2022-07-29